extract token streaming into bitnet-token-stream microcrate
